third_party/WebKit/LayoutTests/fast/dom/shadow/content-pseudo-element-with-hat-expected.txt

Test for ::content with /shadow-all/ and :host.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.getComputedStyle(document.querySelector("#host1 > .content")).color is "rgb(0, 128, 0)"
PASS window.getComputedStyle(document.querySelector("#host1 > .content")).color is "rgb(0, 128, 0)"
PASS window.getComputedStyle(document.querySelector("#host1 > .content")).color is not "rgb(255, 0, 0)"
PASS window.getComputedStyle(document.querySelector("#host1 > .content")).color is not "rgb(255, 0, 0)"
PASS successfullyParsed is true

TEST COMPLETE
